+++ Core/Utilities/Service/ServiceHostResolverExtensions.cs
@@ -82,7 +82,7 @@ namespace Unity.Cloud.Common
             if (domainResolverOverride == null)
                 throw new ArgumentNullException(nameof(domainResolverOverride));
 
-            return new ServiceHostResolver(serviceHostResolver.GetResolvedServiceHost(), domainResolverOverride);
+            return new ServiceHostResolver(serviceHostResolver, domainResolverOverride);
         }
 
         /// <summary>