+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2025-12-09
+
+### Changed
+- `DotNetHttpClient` now supports automatic decompression for `Deflate` and `GZip`.
+- `ServiceHostResolver` supports staging cloud environment override from Unity Editor launch argument.
+
+### Fixed
+- Fixed `ServiceHostResolver` internal copy not copying the original service host options.
+
 ## [1.5.0] - 2025-10-15
 
 ### Changed